splits covariance correlation test into separate tests
